backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.6
PyJWT==2.8.0
